added the vte junk causer
Older versions of VTE did not have the ability to ignore unknown
escape sequences, leading to junk in the terminal as they
are printed to the screen.

wal/pywal have always included a command-line switch to disable
this sequence. pywal in addition uses horrible escape sequence
magic to work around the issue automatically.

This sequence controls /border/ background color (padding area)
in many terminal emulators. VTE doesn't require it and thus
doesn't support it.

Newer versions of VTE DO have the ability to ignore unknown
escape sequences and we can finally just print the sequence to
the terminal.

Your VTE is too old, update it.

https://bugzilla.gnome.org/show_bug.cgi?id=403130
https://github.com/dylanaraps/pywal/issues/69
https://gitlab.gnome.org/GNOME/vte/-/issues/52
http://pod.tst.eu/http://cvs.schmorp.de/rxvt-unicode/doc/rxvt.7.pod#XTerm_Operating_System_Commands